+++ postgres-types/src/time_03.rs
@@ -30,7 +30,6 @@ impl ToSql for PrimitiveDateTime {
     }
 
     accepts!(TIMESTAMP);
-    to_sql_checked!();
 }
 
 impl<'a> FromSql<'a> for OffsetDateTime {
@@ -56,7 +55,6 @@ impl ToSql for OffsetDateTime {
     }
 
     accepts!(TIMESTAMPTZ);
-    to_sql_checked!();
 }
 
 impl<'a> FromSql<'a> for Date {
@@ -80,7 +78,6 @@ impl ToSql for Date {
     }
 
     accepts!(DATE);
-    to_sql_checked!();
 }
 
 impl<'a> FromSql<'a> for Time {
@@ -104,5 +101,4 @@ impl ToSql for Time {
     }
 
     accepts!(TIME);
-    to_sql_checked!();
 }